dbus => 1.15.4, expat => 2.5.0 (#8079)
* dbus => 1.15.4

* expat => 2.5.0

* rebuild dbus

* add libtool archive for imagemagick install

+++ packages/dbus.rb
@@ -3,59 +3,59 @@ require 'package'
 class Dbus < Package
   description 'D-Bus is a message bus system, a simple way for applications to talk to one another.'
   homepage 'https://www.freedesktop.org/wiki/Software/dbus/'
-  version '1.13.18-1'
+  version '1.15.4'
   license 'Apache-2.0'
   compatibility 'all'
-  source_url 'https://dbus.freedesktop.org/releases/dbus/dbus-1.13.18.tar.xz'
-  source_sha256 '8078f5c25e34ab907ce06905d969dc8ef0ccbec367e1e1707c7ecf8460f4254e'
+  source_url 'https://gitlab.freedesktop.org/dbus/dbus.git'
+  git_hashtag "dbus-#{version}"
 
   binary_url({
-    aarch64: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.13.18-1_armv7l/dbus-1.13.18-1-chromeos-armv7l.tpxz',
-     armv7l: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.13.18-1_armv7l/dbus-1.13.18-1-chromeos-armv7l.tpxz',
-       i686: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.13.18-1_i686/dbus-1.13.18-1-chromeos-i686.tpxz',
-     x86_64: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.13.18-1_x86_64/dbus-1.13.18-1-chromeos-x86_64.tpxz'
+    aarch64: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.15.4_armv7l/dbus-1.15.4-chromeos-armv7l.tar.zst',
+     armv7l: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.15.4_armv7l/dbus-1.15.4-chromeos-armv7l.tar.zst',
+       i686: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.15.4_i686/dbus-1.15.4-chromeos-i686.tar.zst',
+     x86_64: 'https://gitlab.com/api/v4/projects/26210301/packages/generic/dbus/1.15.4_x86_64/dbus-1.15.4-chromeos-x86_64.tar.zst'
   })
   binary_sha256({
-    aarch64: '85709ef07ef65ade2689ecbce629200f349f493247b949817f3272e6cb504327',
-     armv7l: '85709ef07ef65ade2689ecbce629200f349f493247b949817f3272e6cb504327',
-       i686: '0315f3eced7198a0850973289f8af06e9f5bdc7aec29834debfaa56556c29c98',
-     x86_64: '3a459724ba6e7f14eacc089fb5500b20b7857aa3034cfc58edf2450ad8ac1c02'
+    aarch64: '3e8cb13e7c1e6bb23067f0def5eb33c1fb042b99eac3654a2a02a64cd52d5f59',
+     armv7l: '3e8cb13e7c1e6bb23067f0def5eb33c1fb042b99eac3654a2a02a64cd52d5f59',
+       i686: 'd48d4149c92bef9981515ce47a4491123013b855147290f358126e18d1e5d4b7',
+     x86_64: 'fc1c818d773ece8193082dca1203878e8382a3fb17d7c6cd32a1f29d9f4788e4'
   })
 
-  depends_on 'expat' => :build
   depends_on 'libice' => :build
   depends_on 'libsm' => :build
-  depends_on 'libx11' => :build
-
-  def self.patch
-    system 'filefix'
-  end
+  depends_on 'expat' # R
+  depends_on 'gcc' # R
+  depends_on 'glibc' # R
+  depends_on 'libx11' # R
 
   def self.build
-    system "#{CREW_ENV_OPTIONS} \
-              ./configure #{CREW_OPTIONS} \
-              --disable-doxygen-docs \
-              --disable-xml-docs \
-              --disable-apparmor \
-              --disable-libaudit \
-              --enable-x11-autolaunch \
-              --localstatedir=/var \
-              --runstatedir=/var/run \
-              --with-dbus-session-bus-listen-address=unix:tmpdir=/tmp"
-    system 'make'
+    system "mold -run meson setup #{CREW_MESON_OPTIONS} \
+      -Dapparmor=disabled \
+      -Ddbus_session_bus_listen_address=unix:tmpdir=/tmp \
+      -Ddoxygen_docs=disabled \
+      -Dlibaudit=disabled \
+      -Dlaunchd=disabled \
+      -Dxml_docs=disabled \
+      -Druntime_dir=/var \
+      -Dsystemd=disabled \
+      -Dx11_autolaunch=enabled \
+       builddir"
+    system 'meson configure builddir'
+    system "mold -run #{CREW_NINJA} -C builddir"
   end
 
   def self.install
-    system 'make', "DESTDIR=#{CREW_DEST_DIR}", 'install'
+    system "DESTDIR=#{CREW_DEST_DIR} #{CREW_NINJA} -C builddir install"
     FileUtils.mkdir_p "#{CREW_DEST_PREFIX}/etc/env.d/"
     @dbusconfigenv = <<~DBUSCONFIGEOF
       # Dbus settings
       DBUS_SYSTEM_BUS_ADDRESS='unix:path=/var/run/dbus/system_bus_socket'
-      [[ "\$DBUS_SESSION_BUS_ADDRESS"=="disabled:" ]] && unset DBUS_SESSION_BUS_ADDRESS
-      if [ -z "\${DBUS_SESSION_BUS_ADDRESS}" ]; then
+      [[ "$DBUS_SESSION_BUS_ADDRESS"=="disabled:" ]] && unset DBUS_SESSION_BUS_ADDRESS
+      if [ -z "${DBUS_SESSION_BUS_ADDRESS}" ]; then
         # if not found, launch a new one
         eval `dbus-launch --sh-syntax`
-        # echo "D-Bus per-session daemon address is: \$DBUS_SESSION_BUS_ADDRESS"
+        # echo "D-Bus per-session daemon address is: $DBUS_SESSION_BUS_ADDRESS"
       fi
       dbus-update-activation-environment --all
     DBUSCONFIGEOF
@@ -63,6 +63,6 @@ class Dbus < Package
   end
 
   def self.postinstall
-    FileUtils.rm_f "#{CREW_PREFIX}/etc/env.d/03-dbus"
+    FileUtils.rm_f "#{CREW_PREFIX}/etc/env.d/03-dbus" if File.file?("#{CREW_PREFIX}/etc/env.d/03-dbus")
   end
 end
